Implement API Consuming in /transactions/report
- Changed service input datatypes to make it more easily usable and added "token".
- Fixed dates in models once more
- Primitive "int" causes errors on consuming API(?). Switched all to "Integer".

+++ src/main/java/net/mert/reportingapi/model/TransactionReport.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.Max;
 @AllArgsConstructor
 @NoArgsConstructor
 public class TransactionReport {
-    private int count, total;
+    private Long count, total;
 
     @Max(3)
     private String currency;
